Fix non-adherence to normalisation standard.
The code that normalises passphrases prior to their use was not
returning the expected value in some situations.  This addresses that
issue, ensuring that the value returned matches the expectation.

Because pre-encrypted data will not decrypt correctly with the new
mechanism the decryption function now attempts to use the old-style
normalisation code if the first attempt fails.  This increases the time
to return an error if the passphrase is incorrect, but ensures that
valid passphrases will decrypt the data.